+++ src/main/java/com/felmero/feladat/controller/PostController.java
@@ -1,13 +1,11 @@
 package com.felmero.feladat.controller;
 
-import org.aspectj.apache.bcel.generic.ReturnaddressType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -23,31 +21,28 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor(onConstructor = @__({ @Autowired }))
 public class PostController {
 
-	
 	private final PostService postService;
-	
-	
+
 	@PostMapping
 	@ResponseBody
-	public ResponseEntity<?> savePost(@RequestBody Post post){
+	public ResponseEntity<?> savePost(@RequestBody Post post) {
 		try {
 			return ResponseEntity.ok(postService.save(post));
 		} catch (Exception e) {
 			return ResponseEntity.badRequest().build();
 		}
 	}
-	
+
 	@GetMapping("/{id}")
 	@ResponseBody
-	public ResponseEntity<Post> getPost(@PathVariable(value="id") Integer id) {
+	public ResponseEntity<Post> getPost(@PathVariable(value = "id") Long id) {
 		try {
 			Post post = postService.get(id);
 			return ResponseEntity.ok(post);
 		} catch (Exception e) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-			
 
 		}
 	}
-	
+
 }
